Initial updates for converting the management console to use a relational database and spring data jpa.

+++ account-management-app/src/main/java/org/duracloud/account/app/controller/AccountEditForm.java
@@ -16,7 +16,7 @@ public class AccountEditForm {
 
     private String department;
     
-    private int accountClusterId;
+    private Long accountClusterId;
 
     @NotBlank(message = "You must specify an account name.")
     private String acctName;
@@ -45,11 +45,11 @@ public class AccountEditForm {
         this.department = department;
     }
 
-    public int getAccountClusterId() {
+    public Long getAccountClusterId() {
         return accountClusterId;
     }
 
-    public void setAccountClusterId(int accountClusterId) {
+    public void setAccountClusterId(Long accountClusterId) {
         this.accountClusterId = accountClusterId;
     }
