+++ account-management-app/src/main/java/org/duracloud/account/app/controller/AccountSetupForm.java
@@ -1,14 +1,13 @@
 package org.duracloud.account.app.controller;
 
-import java.util.LinkedList;
-import java.util.List;
-
-import javax.validation.Valid;
-
-import org.duracloud.account.common.domain.ComputeProviderAccount;
-import org.duracloud.account.common.domain.StorageProviderAccount;
+import org.duracloud.account.db.model.ComputeProviderAccount;
+import org.duracloud.account.db.model.StorageProviderAccount;
 import org.duracloud.storage.domain.StorageProviderType;
 import org.hibernate.validator.constraints.NotBlank;
+
+import javax.validation.Valid;
+import java.util.LinkedList;
+import java.util.List;
 
 public class AccountSetupForm {
     
@@ -134,7 +133,7 @@ public class AccountSetupForm {
     }
 
     public static class StorageCredentials {
-        private int id;
+        private Long id;
         @NotBlank(message="Username is required")
         private String username;
         @NotBlank(message="Password is required")
@@ -158,10 +157,10 @@ public class AccountSetupForm {
         
         public StorageCredentials() {}
 
-        public int getId() {
+        public Long getId() {
             return id;
         }
-        public void setId(int id) {
+        public void setId(Long id) {
             this.id = id;
         }
