+++ account-management-app/src/main/java/org/duracloud/account/app/controller/NewAccountForm.java
@@ -3,21 +3,19 @@
  */
 package org.duracloud.account.app.controller;
 
-import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.LinkedList;
-import java.util.List;
-
 import org.duracloud.account.annotation.UniqueSubdomainConstraint;
-import org.duracloud.account.common.domain.AccountType;
-import org.duracloud.account.common.domain.ServicePlan;
-import org.duracloud.storage.domain.StorageProviderType;
+import org.duracloud.account.db.model.AccountType;
+import org.duracloud.account.db.model.ServicePlan;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.NotBlank;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 import org.springframework.web.context.WebApplicationContext;
+
+import java.io.Serializable;
+import java.util.Arrays;
+import java.util.LinkedList;
+import java.util.List;
 
 /**
  * @author Daniel Bernstein
@@ -43,7 +41,7 @@ public class NewAccountForm implements Serializable {
 
     private AccountType accountType;
 
-    private Integer accountClusterId;
+    private Long accountClusterId;
 
     public List<ServicePlan> getServicePlans(){
         return new LinkedList<ServicePlan>(Arrays.asList(ServicePlan.values()));
@@ -96,11 +94,11 @@ public class NewAccountForm implements Serializable {
         this.accountType = accountType;
     }
 
-    public Integer getAccountClusterId() {
+    public Long getAccountClusterId() {
         return accountClusterId;
     }
 
-    public void setAccountClusterId(Integer accountClusterId) {
+    public void setAccountClusterId(Long accountClusterId) {
         this.accountClusterId = accountClusterId;
     }
